[TS migration] Migrate 'useDragAndDrop.js' hook and 'withEnvironment.js' HOC
