--- src/hooks/useEnvironment.js
@@ -1,13 +1,0 @@
-import {useContext} from 'react';
-import CONST from '../CONST';
-import {EnvironmentContext} from '../components/withEnvironment';
-
-export default function useEnvironment() {
-    const {environment, environmentURL} = useContext(EnvironmentContext);
-    return {
-        environment,
-        environmentURL,
-        isProduction: environment === CONST.ENVIRONMENT.PRODUCTION,
-        isDevelopment: environment === CONST.ENVIRONMENT.DEV,
-    };
-}
