+++ src/pages/ShareCodePage.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {View, ScrollView} from 'react-native';
 import _ from 'underscore';
+import PropTypes from 'prop-types';
 import ScreenWrapper from '../components/ScreenWrapper';
 import HeaderWithBackButton from '../components/HeaderWithBackButton';
 import Navigation from '../libs/Navigation/Navigation';
@@ -20,16 +21,18 @@ import CONST from '../CONST';
 import ContextMenuItem from '../components/ContextMenuItem';
 import * as UserUtils from '../libs/UserUtils';
 import ROUTES from '../ROUTES';
-import withEnvironment, {environmentPropTypes} from '../components/withEnvironment';
+import withEnvironment from '../components/withEnvironment';
 import * as Url from '../libs/Url';
 
 const propTypes = {
     /** The report currently being looked at */
     report: reportPropTypes,
 
+    /** The string value representing the URL of the current environment */
+    environmentURL: PropTypes.string.isRequired,
+
     ...withLocalizePropTypes,
     ...withCurrentUserPersonalDetailsPropTypes,
-    ...environmentPropTypes,
 };
 
 const defaultProps = {
